Add car leaving handler and restructure
- when a car leaves a parking space, gate will handle it
- moves token related class to another package
