+++ gate/src/main/java/edu/rutgers/cs431/teamchen/gate/token/PeerHttpAddressProvider.java
@@ -1,4 +1,4 @@
-package edu.rutgers.cs431.teamchen.gate;
+package edu.rutgers.cs431.teamchen.gate.token;
 
 import java.net.URL;
 import java.util.ArrayList;
