+++ gate/src/main/java/edu/rutgers/cs431/teamchen/gate/token/TokenStore.java
@@ -1,4 +1,4 @@
-package edu.rutgers.cs431.teamchen.gate;
+package edu.rutgers.cs431.teamchen.gate.token;
 
 public interface TokenStore {
     // returns a token, if there is no token then wait indefinitely
